All UI scenes are now organized in folders ("HUD", "Menu", and "Tutorial"). The scene used for the first release is now under "Scene" folder.
Added scene for test purposes (Test_Game_Zone), which is a copy of ES2015A, with an extra GameObject to be able to instantiate the player's stronghold.

Can now instantiate one civilization for the player.

IMPORTANT: Main_Game.cs is now under a new folder (Gameplay) and is responsible for instantiating the HUD and the player's stronghold, however, git assumes the original file was deleted and not moved.

There's a new class (GameInformation.cs) used in the scene UI_Select_Civilization to hold all the information to be used later on within Main_Game to instantiate the settings selected by the user. This is done to separate the UI logic of GameSetupScript.cs from the actual information that is going to be used.