Detects the end of the game. Tested when the user wins and loses.
The AI's MicroManager now detects the end of the game to avoid errors (so it's not only the AIController).

Unregisters the events subscribed in ResourcesPlacer. PauseMenuLogic no longer handles the call to clear the other subscribed events from 0e432201ec6582001993b54ee6e3884f48be38e5. They are handled with OnDestroy().